Make Sinatra middleware better
The Sinatra middleware is now reasonably well tested. I also added a
simple way to manage the development dependencies for working with the
middleware specifically. I don't _exactly_ like it, because it means
that when you bundle install Loginator, you get the whole middleware
chain too, but I'll figure out a way around that eventually.